Move observations API classes from wastd/api.py to observations
